requirements.txt
# Create and activate virtual environment
python -m venv cookenv
# On Windows: cookenv\Scripts\activate
# On Mac/Linux: source cookenv/bin/activate

# Create requirements.txt
echo "streamlit==1.28.0
pillow==10.1.0
opencv-python==4.8.1
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.0
googletrans==4.0.0-rc1" > requirements.txt

# Install
